+++ app/src/main/java/org/gampiot/robokide/ui/fragments/editor/EditorFragment.kt
@@ -15,7 +15,7 @@ import com.google.android.material.tabs.TabLayout
 
 import org.gampiot.robokide.R
 import org.gampiot.robokide.databinding.FragmentEditorBinding
-import org.gampiot.robokide.feature.base.RobokFragment
+import org.gampiot.robokide.feature.base.ui.RobokFragment
 import org.gampiot.robokide.feature.component.terminal.RobokTerminal
 import org.gampiot.robokide.feature.component.progress.DotProgressBar
 import org.gampiot.robokide.ui.fragments.build.output.OutputFragment
